changing pages redirectory
