+++ front/components/home/content/Product/VerticalSection.tsx
@@ -10,20 +10,26 @@ import { H2, P } from "@app/components/home/ContentComponents";
 import { CustomerCaroussel } from "@app/pages/home/solutions/customer-support";
 import { DataCaroussel } from "@app/pages/home/solutions/data-analytics";
 import { EngineeringCaroussel } from "@app/pages/home/solutions/engineering";
+import { ITCaroussel } from "@app/pages/home/solutions/it";
 import { KnowledgeCaroussel } from "@app/pages/home/solutions/knowledge";
+import { LegalCaroussel } from "@app/pages/home/solutions/legal";
 import { MarketingCaroussel } from "@app/pages/home/solutions/marketing";
-import { RecruitingCaroussel } from "@app/pages/home/solutions/recruiting-people";
+import { ProductivityCaroussel } from "@app/pages/home/solutions/productivity";
+import { PeopleCaroussel } from "@app/pages/home/solutions/recruiting-people";
 import { SalesCaroussel } from "@app/pages/home/solutions/sales";
 
 export function VerticalSection() {
   const carouselSections = [
-    { title: "Customer Support", component: CustomerCaroussel },
     { title: "Sales", component: SalesCaroussel },
-    { title: "Marketing", component: MarketingCaroussel },
-    { title: "Recruiting", component: RecruitingCaroussel },
+    { title: "Customer Support", component: CustomerCaroussel },
     { title: "Engineering", component: EngineeringCaroussel },
-    { title: "Knowledge", component: KnowledgeCaroussel },
     { title: "Data Analytics", component: DataCaroussel },
+    { title: "IT", component: ITCaroussel },
+    { title: "Knowledge", component: KnowledgeCaroussel },
+    { title: "Legal", component: LegalCaroussel },
+    { title: "Marketing", component: MarketingCaroussel },
+    { title: "People", component: PeopleCaroussel },
+    { title: "Productivity", component: ProductivityCaroussel },
   ];
 
   const [api, setApi] = React.useState<any>(null);
