+++ front/pages/home/solutions/data-analytics.tsx
@@ -1,14 +1,22 @@
+import { Button } from "@dust-tt/sparkle";
+import Link from "next/link";
 import type { ReactElement } from "react";
-import React from "react";
 
-import { BlogSection } from "@app/components/home/content/Product/BlogSection";
+import { BenefitsSection } from "@app/components/home/content/Solutions/BenefitsSection";
 import {
-  pageSettings as carouselSettings,
+  Benefits,
+  DemoVideo,
+  Hero,
+  pageSettings,
   Quote,
+  UseCases,
 } from "@app/components/home/content/Solutions/configs/dataConfig";
+import { DemoVideoSection } from "@app/components/home/content/Solutions/DemoVideoSection";
+import { HeroSection } from "@app/components/home/content/Solutions/HeroSection";
+import { UseCasesSection } from "@app/components/home/content/Solutions/UseCasesSection";
 import {
   CarousselContentBlock,
-  HeaderContentBlock,
+  QuoteSection,
 } from "@app/components/home/ContentBlocks";
 import { Grid } from "@app/components/home/ContentComponents";
 import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
@@ -17,212 +25,112 @@ import {
   getParticleShapeIndexByName,
   shapeNames,
 } from "@app/components/home/Particles";
-import type { SolutionSectionAssistantBlockProps } from "@app/components/home/SolutionSection";
-import { SolutionSection } from "@app/components/home/SolutionSection";
 import TrustedBy from "@app/components/home/TrustedBy";
+import { classNames } from "@app/lib/utils";
 
 export async function getServerSideProps() {
   return {
     props: {
-      shape: getParticleShapeIndexByName(shapeNames.wave),
+      shape: getParticleShapeIndexByName(shapeNames.octahedron),
       gtmTrackingId: process.env.NEXT_PUBLIC_GTM_TRACKING_ID ?? null,
     },
   };
 }
 
-interface pageSettingsProps {
-  uptitle: string;
-  title: React.ReactNode;
-  description: React.ReactNode;
-  from: string;
-  to: string;
-}
+const GRID_SECTION_CLASSES = classNames(
+  "flex flex-col gap-8",
+  "col-span-12",
+  "lg:col-span-12 lg:col-start-1",
+  "xl:col-span-12 xl:col-start-1",
+  "2xl:col-start-1"
+);
 
-const pageSettings: pageSettingsProps = {
-  uptitle: "Data and Analytics",
-  title: <>From Data to&nbsp;Action</>,
-  from: "from-amber-200",
-  to: "to-amber-500",
-  description: (
-    <>
-      Focus on&nbsp;generating first-of-a-kind insights while AI helps with
-      routine queries and&nbsp;standard reports across teams.
-    </>
-  ),
-};
-
-export default function DataAnalytics() {
+export default function Data() {
   return (
     <>
-      <HeaderContentBlock
-        uptitle={"Dust for " + pageSettings.uptitle}
-        title={pageSettings.title}
-        from={pageSettings.from}
-        to={pageSettings.to}
-        subtitle={pageSettings.description}
-      />
-      <TrustedBy />
-      <Grid>
-        <SolutionSection
-          title={
-            <>
-              Give everyone access
-              <br />
-              to&nbsp;data analysis.
-            </>
-          }
-          blocks={[
-            {
-              color: "amber",
-              contentBlocks: [
-                {
-                  title: <>Democratize data analysis</>,
-                  content: [
-                    <>
-                      Enable anyone at the company to generate recurring
-                      analyses using natural language to create SQL queries that
-                      understand your database schemas.
-                    </>,
-                  ],
-                },
-                {
-                  title: <>Reduce time to data insights</>,
-                  content: [
-                    <>
-                      Turn .csv files, Notion databases, and Google Sheets into
-                      data sources for quantitative analyses with agents that
-                      have a built-in understanding of your company’s business
-                      terminology.
-                    </>,
-                  ],
-                },
-              ],
-
-              assistantBlocks: [assistantExamples[0], assistantExamples[1]],
-            },
-          ]}
+      <div className="container flex w-full flex-col gap-0 px-2 py-2 pb-12">
+        <HeroSection
+          {...Hero}
+          fromColor={pageSettings.from}
+          toColor={pageSettings.to}
         />
-        <SolutionSection
-          title={
-            <>
-              Free the&nbsp;team
-              <br />
-              from&nbsp;being a&nbsp;perpetual help&nbsp;desk.
-            </>
-          }
-          blocks={[
-            {
-              color: "amber",
-              contentBlocks: [
-                {
-                  title: (
-                    <>
-                      Streamline onboarding
-                      <br />
-                      for the&nbsp;data&nbsp;team
-                    </>
-                  ),
-                  content: [
-                    <>
-                      Share up-to-date runbooks and internal documentation for
-                      new data team members to ramp up efficiently and
-                      autonomously with conversational agents.
-                    </>,
-                  ],
-                },
-                {
-                  title: (
-                    <>
-                      Improve internal documentation on&nbsp;internal
-                      data&nbsp;sources
-                    </>
-                  ),
-                  content: [
-                    <>
-                      Clean up or draft great documentation about your
-                      infrastructure or fields, tables and relationships to be
-                      clear to all.
-                    </>,
-                  ],
-                },
-              ],
-              assistantBlocks: [assistantExamples[2], assistantExamples[3]],
-            },
-          ]}
-        />
-      </Grid>
-      <BlogSection
-        headerColorFrom="from-amber-200"
-        headerColorTo="to-amber-500"
-      />
+        <Grid>
+          <div className={GRID_SECTION_CLASSES}>
+            <BenefitsSection
+              benefits={Benefits}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+            {/* <MetricSection {...Metrics} /> */}
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <UseCasesSection
+              useCase={UseCases}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <DemoVideoSection
+              demoVideo={DemoVideo}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <QuoteSection {...Quote} />
+            {/* <CustomerStoriesSection
+              title="Customer stories"
+              stories={Stories}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            /> */}
+          </div>
+          <TrustedBy />
+          <div className={GRID_SECTION_CLASSES}>
+            {Hero.ctaButtons && (
+              <div className="mt-4 flex justify-center gap-4">
+                {Hero.ctaButtons.primary && (
+                  <Link href={Hero.ctaButtons.primary.href} shallow={true}>
+                    <Button
+                      variant="highlight"
+                      size="md"
+                      label={Hero.ctaButtons.primary.label}
+                      icon={Hero.ctaButtons.primary.icon}
+                    />
+                  </Link>
+                )}
+                {Hero.ctaButtons.secondary && (
+                  <Button
+                    variant="outline"
+                    size="md"
+                    label={Hero.ctaButtons.secondary.label}
+                    href={Hero.ctaButtons.secondary.href}
+                  />
+                )}
+              </div>
+            )}
+          </div>
+        </Grid>
+      </div>
     </>
   );
 }
 
-DataAnalytics.getLayout = (
-  page: ReactElement,
-  pageProps: LandingLayoutProps
-) => {
+Data.getLayout = (page: ReactElement, pageProps: LandingLayoutProps) => {
   return <LandingLayout pageProps={pageProps}>{page}</LandingLayout>;
 };
-
-const assistantExamples: SolutionSectionAssistantBlockProps[] = [
-  {
-    emoji: "💬",
-    name: "@SQLbuddy",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Generates simple SQL queries that understand business logic
-        and&nbsp;fixes or&nbsp;improves existing ones
-      </>
-    ),
-  },
-  {
-    emoji: "🔬",
-    name: "@analyst",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Answers advanced data questions on your behalf by writing and executing
-        queries and providing visualizations.
-      </>
-    ),
-  },
-  {
-    emoji: "📈",
-    name: "@dataRun",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Answers questions about internal processes and&nbsp;runbooks on&nbsp;the
-        data team
-      </>
-    ),
-  },
-  {
-    emoji: "📚",
-    name: "@dataModel",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Knows everything about data schemas and&nbsp;data model relationships
-        at&nbsp;the company
-      </>
-    ),
-  },
-];
 
 export function DataCaroussel() {
   return (
     <CarousselContentBlock
-      title={carouselSettings.uptitle}
-      from={carouselSettings.from}
-      to={carouselSettings.to}
+      title={pageSettings.uptitle}
+      from={pageSettings.from}
+      to={pageSettings.to}
       border="border-pink-100/60"
       href="/home/solutions/data-analytics"
-      bulletPoints={carouselSettings.bulletPoints}
-      image={carouselSettings.image}
+      bulletPoints={pageSettings.bulletPoints}
+      image={pageSettings.image}
       quote={Quote}
     />
   );
