+++ front/pages/home/solutions/engineering.tsx
@@ -1,14 +1,26 @@
+import { Button } from "@dust-tt/sparkle";
+import Link from "next/link";
 import type { ReactElement } from "react";
-import React from "react";
 
-import { BlogSection } from "@app/components/home/content/Product/BlogSection";
+import { BenefitsSection } from "@app/components/home/content/Solutions/BenefitsSection";
 import {
-  pageSettings as carouselSettings,
+  Benefits,
+  DemoVideo,
+  Hero,
+  Metrics,
+  pageSettings,
   Quote,
+  Stories,
+  UseCases,
 } from "@app/components/home/content/Solutions/configs/engineeringConfig";
+import { CustomerStoriesSection } from "@app/components/home/content/Solutions/CustomerStoriesSection";
+import { DemoVideoSection } from "@app/components/home/content/Solutions/DemoVideoSection";
+import { HeroSection } from "@app/components/home/content/Solutions/HeroSection";
+import { UseCasesSection } from "@app/components/home/content/Solutions/UseCasesSection";
 import {
   CarousselContentBlock,
-  HeaderContentBlock,
+  MetricSection,
+  QuoteSection,
 } from "@app/components/home/ContentBlocks";
 import { Grid } from "@app/components/home/ContentComponents";
 import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
@@ -17,162 +29,94 @@ import {
   getParticleShapeIndexByName,
   shapeNames,
 } from "@app/components/home/Particles";
-import type { SolutionSectionAssistantBlockProps } from "@app/components/home/SolutionSection";
-import { SolutionSection } from "@app/components/home/SolutionSection";
 import TrustedBy from "@app/components/home/TrustedBy";
+import { classNames } from "@app/lib/utils";
 
 export async function getServerSideProps() {
   return {
     props: {
-      shape: getParticleShapeIndexByName(shapeNames.cube),
+      shape: getParticleShapeIndexByName(shapeNames.octahedron),
       gtmTrackingId: process.env.NEXT_PUBLIC_GTM_TRACKING_ID ?? null,
     },
   };
 }
 
-interface pageSettingsProps {
-  uptitle: string;
-  title: React.ReactNode;
-  description: React.ReactNode;
-  from: string;
-  to: string;
-}
-
-const pageSettings: pageSettingsProps = {
-  uptitle: "Engineers and Developers",
-  title: (
-    <>
-      Code smarter,
-      <br />
-      Resolve faster
-    </>
-  ),
-  from: "from-emerald-200",
-  to: "to-emerald-500",
-  description: (
-    <>
-      Reduce interruptions, write better code, speed up incident response,
-      and&nbsp;accelerate new engineers' onboarding.
-    </>
-  ),
-};
+const GRID_SECTION_CLASSES = classNames(
+  "flex flex-col gap-8",
+  "col-span-12",
+  "lg:col-span-12 lg:col-start-1",
+  "xl:col-span-12 xl:col-start-1",
+  "2xl:col-start-1"
+);
 
 export default function Engineering() {
   return (
     <>
-      <HeaderContentBlock
-        uptitle={"Dust for " + pageSettings.uptitle}
-        title={pageSettings.title}
-        from={pageSettings.from}
-        to={pageSettings.to}
-        subtitle={pageSettings.description}
-      />
-      <TrustedBy />
-      <Grid>
-        <SolutionSection
-          title={"Improve Code Quality."}
-          blocks={[
-            {
-              color: "emerald",
-              contentBlocks: [
-                {
-                  title: <>Talk to&nbsp;your&nbsp;codebase</>,
-                  content: [
-                    <>
-                      Get conversational access to&nbsp;your entire code base.
-                    </>,
-                    <>
-                      Accelerate onboarding for new engineers on&nbsp;the team
-                      and&nbsp;reduce interruptions from&nbsp;other teams.
-                    </>,
-                  ],
-                },
-                {
-                  title: <>Create your&nbsp;architecture copilot.</>,
-                  content: (
-                    <>
-                      Specialize the&nbsp;best models (Gemini, GPT4, Mistral)
-                      to&nbsp;answer general code questions with concise,
-                      straight-to-the-point answers that have context
-                      on&nbsp;the team's stack, runbooks, and&nbsp;architecture
-                      preferences.
-                    </>
-                  ),
-                },
-              ],
-              assistantBlocks: [
-                assistantExamples[0],
-                assistantExamples[1],
-                assistantExamples[5],
-              ],
-            },
-          ]}
+      <div className="container flex w-full flex-col gap-0 px-2 py-2 pb-12">
+        <HeroSection
+          {...Hero}
+          fromColor={pageSettings.from}
+          toColor={pageSettings.to}
         />
-        <SolutionSection
-          title={
-            <>
-              Fight fires,
-              <br />
-              get&nbsp;back to&nbsp;Coding.
-            </>
-          }
-          blocks={[
-            {
-              color: "emerald",
-              contentBlocks: [
-                {
-                  title: (
-                    <>
-                      Protect your&nbsp;eng team with a&nbsp;help desk for most
-                      situations
-                    </>
-                  ),
-                  content: [
-                    <>
-                      Provide answers to&nbsp;questions from&nbsp;the rest
-                      of&nbsp;the company automatically to&nbsp;avoid
-                      interruptions.
-                    </>,
-                    <>
-                      Give developers context on&nbsp;previous incidents
-                      in&nbsp;seconds, without the&nbsp;need to&nbsp;ping last
-                      time's hero.
-                    </>,
-                  ],
-                },
-                {
-                  title: (
-                    <>
-                      Get the&nbsp;report done for the&nbsp;company to&nbsp;stay
-                      in&nbsp;the know
-                    </>
-                  ),
-                  content: [
-                    <>
-                      Generate weekly summaries on&nbsp;what shipped, what
-                      broke, and&nbsp;what's in&nbsp;flight.
-                    </>,
-                    <>
-                      Make these available to&nbsp;the team's stakeholders with
-                      explanations on&nbsp;the technical terms they might not
-                      understand.
-                    </>,
-                  ],
-                },
-              ],
-              assistantBlocks: [
-                assistantExamples[3],
-                assistantExamples[4],
-                assistantExamples[2],
-              ],
-            },
-          ]}
-        />
-      </Grid>
-      <BlogSection
-        headerColorFrom="from-emerald-200"
-        headerColorTo="to-emerald-500"
-      />
+        <Grid>
+          <div className={GRID_SECTION_CLASSES}>
+            <BenefitsSection
+              benefits={Benefits}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+            <MetricSection {...Metrics} />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <UseCasesSection
+              useCase={UseCases}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <DemoVideoSection
+              demoVideo={DemoVideo}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <QuoteSection {...Quote} />
+            <CustomerStoriesSection
+              title="Customer stories"
+              stories={Stories}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <TrustedBy />
+          <div className={GRID_SECTION_CLASSES}>
+            {Hero.ctaButtons && (
+              <div className="mt-4 flex justify-center gap-4">
+                {Hero.ctaButtons.primary && (
+                  <Link href={Hero.ctaButtons.primary.href} shallow={true}>
+                    <Button
+                      variant="highlight"
+                      size="md"
+                      label={Hero.ctaButtons.primary.label}
+                      icon={Hero.ctaButtons.primary.icon}
+                    />
+                  </Link>
+                )}
+                {Hero.ctaButtons.secondary && (
+                  <Button
+                    variant="outline"
+                    size="md"
+                    label={Hero.ctaButtons.secondary.label}
+                    href={Hero.ctaButtons.secondary.href}
+                  />
+                )}
+              </div>
+            )}
+          </div>
+        </Grid>
+      </div>
     </>
   );
 }
@@ -181,82 +125,16 @@ Engineering.getLayout = (page: ReactElement, pageProps: LandingLayoutProps) => {
   return <LandingLayout pageProps={pageProps}>{page}</LandingLayout>;
 };
 
-const assistantExamples: SolutionSectionAssistantBlockProps[] = [
-  {
-    emoji: "📚",
-    name: "@codebase",
-    backgroundColor: "bg-emerald-300",
-    description: <>Answers questions about the&nbsp;company codebase</>,
-  },
-  {
-    emoji: "👨‍💻",
-    name: "@engHelp",
-    backgroundColor: "bg-emerald-300",
-    description: (
-      <>
-        Answers questions from&nbsp;the rest of&nbsp;the company
-        on&nbsp;engineering definitions, ongoing projects, and&nbsp;who's
-        on&nbsp;run
-      </>
-    ),
-  },
-  {
-    emoji: "🚨",
-    name: "@engIncidents",
-    backgroundColor: "bg-emerald-300",
-    description: (
-      <>
-        Assists in&nbsp;the event of&nbsp;an incident with data on&nbsp;previous
-        similar situation and&nbsp;their remediation
-      </>
-    ),
-  },
-  {
-    emoji: "📡",
-    name: "@engWeekly",
-    backgroundColor: "bg-emerald-300",
-    description: (
-      <>
-        Writes a&nbsp;table of&nbsp;shipped and&nbsp;unshipped
-        features—Summarizes incidents with impact, current status,
-        and&nbsp;remediation plans
-      </>
-    ),
-  },
-  {
-    emoji: "⭐️",
-    name: "@engGeneral",
-    backgroundColor: "bg-emerald-300",
-    description: (
-      <>
-        Answers general questions about code architecture and&nbsp;engineering
-        team processes
-      </>
-    ),
-  },
-  {
-    emoji: "🏴‍☠️",
-    name: "@codeGenius",
-    backgroundColor: "bg-emerald-300",
-    description: (
-      <>
-        Answers general questions about code to&nbsp;avoid a&nbsp;trip
-        to&nbsp;StackOverflow
-      </>
-    ),
-  },
-];
-
 export function EngineeringCaroussel() {
   return (
     <CarousselContentBlock
-      title={carouselSettings.uptitle}
-      from={carouselSettings.from}
-      to={carouselSettings.to}
+      title={pageSettings.uptitle}
+      from={pageSettings.from}
+      to={pageSettings.to}
       border="border-pink-100/60"
       href="/home/solutions/engineering"
-      bulletPoints={carouselSettings.bulletPoints}
-      image={carouselSettings.image}
+      bulletPoints={pageSettings.bulletPoints}
+      image={pageSettings.image}
       quote={Quote}
     />
   );
