+++ front/pages/home/solutions/legal.tsx
