+++ front/pages/home/solutions/recruiting-people.tsx
@@ -1,14 +1,26 @@
+import { Button } from "@dust-tt/sparkle";
+import Link from "next/link";
 import type { ReactElement } from "react";
-import React from "react";
 
-import { BlogSection } from "@app/components/home/content/Product/BlogSection";
+import { BenefitsSection } from "@app/components/home/content/Solutions/BenefitsSection";
 import {
-  pageSettings as carouselSettings,
+  Benefits,
+  DemoVideo,
+  Hero,
+  Metrics,
+  pageSettings,
   Quote,
+  Stories,
+  UseCases,
 } from "@app/components/home/content/Solutions/configs/peopleConfig";
+import { CustomerStoriesSection } from "@app/components/home/content/Solutions/CustomerStoriesSection";
+import { DemoVideoSection } from "@app/components/home/content/Solutions/DemoVideoSection";
+import { HeroSection } from "@app/components/home/content/Solutions/HeroSection";
+import { UseCasesSection } from "@app/components/home/content/Solutions/UseCasesSection";
 import {
   CarousselContentBlock,
-  HeaderContentBlock,
+  MetricSection,
+  QuoteSection,
 } from "@app/components/home/ContentBlocks";
 import { Grid } from "@app/components/home/ContentComponents";
 import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
@@ -17,292 +29,112 @@ import {
   getParticleShapeIndexByName,
   shapeNames,
 } from "@app/components/home/Particles";
-import type { SolutionSectionAssistantBlockProps } from "@app/components/home/SolutionSection";
-import { SolutionSection } from "@app/components/home/SolutionSection";
 import TrustedBy from "@app/components/home/TrustedBy";
+import { classNames } from "@app/lib/utils";
 
 export async function getServerSideProps() {
   return {
     props: {
-      shape: getParticleShapeIndexByName(shapeNames.torus),
+      shape: getParticleShapeIndexByName(shapeNames.octahedron),
       gtmTrackingId: process.env.NEXT_PUBLIC_GTM_TRACKING_ID ?? null,
     },
   };
 }
 
-interface pageSettingsProps {
-  uptitle: React.ReactNode;
-  title: React.ReactNode;
-  description: React.ReactNode;
-  from: string;
-  to: string;
-}
+const GRID_SECTION_CLASSES = classNames(
+  "flex flex-col gap-8",
+  "col-span-12",
+  "lg:col-span-12 lg:col-start-1",
+  "xl:col-span-12 xl:col-start-1",
+  "2xl:col-start-1"
+);
 
-const pageSettings: pageSettingsProps = {
-  uptitle: "Recruiting and People",
-  title: <>More&nbsp;time for&nbsp;people, teams, and&nbsp;managers</>,
-  from: "from-amber-200",
-  to: "to-amber-500",
-  description: (
-    <>
-      Support the&nbsp;business and&nbsp;the team effectively across recruiting,
-      onboarding, and&nbsp;career development initiatives.
-    </>
-  ),
-};
-
-export default function RecruitingPeople() {
+export default function People() {
   return (
     <>
-      <HeaderContentBlock
-        uptitle={"Dust for " + pageSettings.uptitle}
-        title={pageSettings.title}
-        from={pageSettings.from}
-        to={pageSettings.to}
-        subtitle={pageSettings.description}
-      />
-      <TrustedBy />
-      <Grid>
-        <SolutionSection
-          title="Focus on&nbsp;being a&nbsp;business partner, not&nbsp;a&nbsp;help desk."
-          blocks={[
-            {
-              color: "amber",
-              contentBlocks: [
-                {
-                  title: (
-                    <>
-                      Give new hires the&nbsp;buddy they can always tap
-                      on&nbsp;the shoulder
-                    </>
-                  ),
-                  content: (
-                    <>
-                      Transform your&nbsp;onboard process with tailored
-                      assistance for newcomers to&nbsp;discover
-                      your&nbsp;methods, people, and&nbsp;culture.
-                    </>
-                  ),
-                },
-                {
-                  title: (
-                    <>Let your&nbsp;internal documentation do&nbsp;the work</>
-                  ),
-                  content: (
-                    <>
-                      Extract yourself from&nbsp;admin ping pong. Allow team
-                      members to&nbsp;interact with an&nbsp;agent that answers
-                      common people-related questions in&nbsp;Slack
-                      and&nbsp;points to&nbsp;the right internal resources.
-                    </>
-                  ),
-                },
-              ],
-              assistantBlocks: [assistantExamples[0], assistantExamples[1]],
-            },
-          ]}
+      <div className="container flex w-full flex-col gap-0 px-2 py-2 pb-12">
+        <HeroSection
+          {...Hero}
+          fromColor={pageSettings.from}
+          toColor={pageSettings.to}
         />
-        <SolutionSection
-          title="Help your&nbsp;People Grow."
-          blocks={[
-            {
-              color: "amber",
-              contentBlocks: [
-                {
-                  title: "Create engaging learning experiences faster",
-                  content: (
-                    <>
-                      Use internal documentation to create training materials
-                      tailored to each role and team.
-                    </>
-                  ),
-                },
-                {
-                  title: "Streamline performance reviews",
-                  content: [
-                    <>
-                      Help reviewers collect data to&nbsp;get a&nbsp;holistic
-                      view of&nbsp;their teammates' impact and&nbsp;make more
-                      accurate evaluations.
-                    </>,
-                    <>
-                      Get more thoughtful peer reviews, with AI-powered feedback
-                      on&nbsp;tone, references to&nbsp;company principles,
-                      priorities and&nbsp;business objectives.
-                    </>,
-                  ],
-                },
-              ],
-              assistantBlocks: [assistantExamples[4], assistantExamples[3]],
-            },
-          ]}
-        />
-        <SolutionSection
-          title="Give your&nbsp;hiring efforts a&nbsp;boost."
-          blocks={[
-            {
-              color: "amber",
-              contentBlocks: [
-                {
-                  title: <>Get a&nbsp;qualified pipeline in&nbsp;a flash</>,
-                  content: [
-                    <>
-                      Easily draft consistent job descriptions. Analyze
-                      and&nbsp;filter candidates' CVs by&nbsp;finding elements
-                      that are a&nbsp;match for expectations.
-                    </>,
-                    <>
-                      Train your&nbsp;team at&nbsp;writing exercises
-                      and&nbsp;questions, reviewing exercises' responses,
-                      and&nbsp;reading through candidates' subtexts.
-                    </>,
-                  ],
-                },
-                {
-                  title: <>Raise the&nbsp;bar on&nbsp;interviews</>,
-                  content: [
-                    <>
-                      Easily prepare the&nbsp;interview process including
-                      questions that match the&nbsp;role expectations
-                      and&nbsp;company guidelines.
-                    </>,
-                    <>
-                      Give the&nbsp;team context on&nbsp;the status of&nbsp;a
-                      hire.
-                    </>,
-                    <>
-                      Help interviewers write consistent
-                      and&nbsp;well-articulated feedback that directly connects
-                      to&nbsp;your company's evaluation rubric.
-                    </>,
-                  ],
-                },
-              ],
-              assistantBlocks: [
-                assistantExamples[2],
-                assistantExamples[5],
-                assistantExamples[7],
-              ],
-            },
-          ]}
-        />
-      </Grid>
-      <BlogSection
-        headerColorFrom="from-amber-200"
-        headerColorTo="to-amber-500"
-      />
+        <Grid>
+          <div className={GRID_SECTION_CLASSES}>
+            <BenefitsSection
+              benefits={Benefits}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+            <MetricSection {...Metrics} />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <UseCasesSection
+              useCase={UseCases}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <DemoVideoSection
+              demoVideo={DemoVideo}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <div className={GRID_SECTION_CLASSES}>
+            <QuoteSection {...Quote} />
+            <CustomerStoriesSection
+              title="Customer stories"
+              stories={Stories}
+              fromColor={pageSettings.from}
+              toColor={pageSettings.to}
+            />
+          </div>
+          <TrustedBy />
+          <div className={GRID_SECTION_CLASSES}>
+            {Hero.ctaButtons && (
+              <div className="mt-4 flex justify-center gap-4">
+                {Hero.ctaButtons.primary && (
+                  <Link href={Hero.ctaButtons.primary.href} shallow={true}>
+                    <Button
+                      variant="highlight"
+                      size="md"
+                      label={Hero.ctaButtons.primary.label}
+                      icon={Hero.ctaButtons.primary.icon}
+                    />
+                  </Link>
+                )}
+                {Hero.ctaButtons.secondary && (
+                  <Button
+                    variant="outline"
+                    size="md"
+                    label={Hero.ctaButtons.secondary.label}
+                    href={Hero.ctaButtons.secondary.href}
+                  />
+                )}
+              </div>
+            )}
+          </div>
+        </Grid>
+      </div>
     </>
   );
 }
 
-RecruitingPeople.getLayout = (
-  page: ReactElement,
-  pageProps: LandingLayoutProps
-) => {
+People.getLayout = (page: ReactElement, pageProps: LandingLayoutProps) => {
   return <LandingLayout pageProps={pageProps}>{page}</LandingLayout>;
 };
 
-const assistantExamples: SolutionSectionAssistantBlockProps[] = [
-  {
-    emoji: "🌱",
-    name: "@onboardingBuddy",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Acts as&nbsp;a friendly guide to&nbsp;help new team members feel welcome
-        and&nbsp;properly informed from&nbsp;day one and&nbsp;as they learn
-        about the&nbsp;company
-      </>
-    ),
-  },
-  {
-    emoji: "👋",
-    name: "@askHR",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Answers questions the&nbsp;People team gets most regularly about company
-        processes and&nbsp;policies based on&nbsp;internal documentation
-        directly on&nbsp;Slack
-      </>
-    ),
-  },
-  {
-    emoji: "🖋️",
-    name: "@hiringOps",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Drafts job descriptions and&nbsp;matching social media communications
-        based on&nbsp;company templates
-      </>
-    ),
-  },
-  {
-    emoji: "🏅",
-    name: "@reviewPrep",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Collects achievements, drafts actionable reviews, helps formulate
-        feedback
-      </>
-    ),
-  },
-  {
-    emoji: "🧑‍🏫",
-    name: "@trainingDraft",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>Designs training modules and&nbsp;crafts employee development plans</>
-    ),
-  },
-  {
-    emoji: "🔬",
-    name: "@screeningPrep",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Provides a&nbsp;suggested analysis of&nbsp;a candidate CV given
-        a&nbsp;role description to&nbsp;highlight questions for a&nbsp;screening
-        interview
-      </>
-    ),
-  },
-  {
-    emoji: "💬",
-    name: "@hiringQuestions",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Suggests interviews questions based on&nbsp;roles, stage of&nbsp;the
-        interview and&nbsp;the company's culture
-      </>
-    ),
-  },
-  {
-    emoji: "🧐",
-    name: "@candidateInfo",
-    backgroundColor: "bg-amber-300",
-    description: (
-      <>
-        Summarize available information about a&nbsp;candidate based
-        on&nbsp;your&nbsp;company database
-      </>
-    ),
-  },
-];
-
-export function RecruitingCaroussel() {
+export function PeopleCaroussel() {
   return (
     <CarousselContentBlock
-      title={carouselSettings.uptitle}
-      from={carouselSettings.from}
-      to={carouselSettings.to}
+      title={pageSettings.uptitle}
+      from={pageSettings.from}
+      to={pageSettings.to}
       border="border-pink-100/60"
       href="/home/solutions/recruiting-people"
-      bulletPoints={carouselSettings.bulletPoints}
-      image={carouselSettings.image}
+      bulletPoints={pageSettings.bulletPoints}
+      image={pageSettings.image}
       quote={Quote}
     />
   );
